+++ src/components/App/Search/Search.scss
@@ -9,7 +9,7 @@
 
 @media only screen and (max-width: 600px) {
   .search {
-    flex-direction: column;
+    flex-direction: row;
   }
 }
