+++ src/containers/Cases/Cases.js
@@ -5,8 +5,6 @@ import Header from '../../components/App/Header/Header'
 import { setup } from '../../services/Auth'
 import Search from '../../components/App/Search/Search'
 
-import SwipeableTemporaryDrawer from '../../components/App/SideBarMenu/sideBarMenu'
-
 const Cases = () => {
   if (setup() === false) {
     return <Redirect to='/' />
@@ -15,7 +13,6 @@ const Cases = () => {
   return (
     <>
       <Header target='home' />
-      <SwipeableTemporaryDrawer />
       <Search />
       <CasesList />
     </>
